+++ engine/src/main/java/dev/gamekit/core/UI.java
@@ -52,9 +52,9 @@ public final class UI {
 
   public UI(Scene scene) {
     Settings settings = Application.getInstance().getSettings();
-    Window win = Window.getInstance();
-    int dw = win.getDisplayWidth();
-    int dh = win.getDisplayHeight();
+    Window.Info windowInfo = Window.getInfo();
+    int dw = windowInfo.displayWidth();
+    int dh = windowInfo.displayHeight();
 
     this.scene = scene;
     this.windowConstraints = new Constraints(dw, dw, dh, dh);
@@ -82,6 +82,10 @@ public final class UI {
 
   void triggerUpdate() {
     needsLayout = true;
+  }
+
+  void clear() {
+    drawTree();
   }
 
   /** Set the initial widget tree */
@@ -357,14 +361,16 @@ public final class UI {
   /** Draws the widget tree to the {@link Window} UI buffer */
   private void drawTree() {
     Window win = Window.getInstance();
+    Window.Info windowInfo = Window.getInfo();
     Graphics2D uiGraphics = win.getUiGraphics();
-    int displayWidth = win.getDisplayWidth();
-    int displayHeight = win.getDisplayHeight();
+    int displayWidth = windowInfo.displayWidth();
+    int displayHeight = windowInfo.displayHeight();
 
     canvasGraphics.setBackground(Constants.TRANSPARENT_COLOR);
     canvasGraphics.clearRect(0, 0, displayWidth, displayHeight);
 
-    tree.render(canvasGraphics);
+    if (tree != null)
+      tree.render(canvasGraphics);
 
     uiGraphics.setBackground(Constants.TRANSPARENT_COLOR);
     uiGraphics.clearRect(0, 0, displayWidth, displayHeight);